#137 move fishing perks to own package
